Make compatibility code for `waitingSince` work on testnet (#2041)
We use to store UNIX timestamps in the `waitingSince` field before
moving to block count. In order to ensure backward compatibility, we
converted from timestamps to blockheight based on the value. Anything
over 1 500 000 was considered a timestamp. But this value is much too
low: on testnet the blockheight is already greater than 2 000 000.

We can use 1 500 000 000 instead, which is somewhere in 2017.

Another way to deal with this would be to simply remove this
compatibility code.