Remove misleading claim-htlc-success log
Whenever a remote commitment is confirmed in which we received htlcs, we
generate a claim-htlc-success transaction even if we don't have the preimage
yet (so we can't broadcast it).

This created a confusing log line:
"tx generation success: desc=claim-htlc-success"

We want to create that log line only when we have the preimage and the node
operator can expect the HTLC to be redeemed on-chain.