Notify node operator on low fee bumping reserve
With anchor outputs, we need to keep utxos available for fee bumping.
Having enough on-chain funds to claim HTLCs in case channels force-close
is necessary to guarantee funds safety.

There is no perfect formula to estimate how much we should reserve: it
depends on the feerate when the channels force-close, the number of
impacted HTLCs and their amount.

We use a very rough estimation, to avoid needlessly spamming node operators
while still notifying them when the situation becomes risky.