Increase timeout onion message integration tests
In channel tests, we set a 60 seconds timeout on the receiver end,
otherwise we sometimes run into timeout on slow CI machines that run a
lot of tests in parallel.

It makes sense to do the same for onion messages, to help us figure out
whether there is a race condition or it's just the machines that are slow.