Avoid default arguments in test helpers
A few of our test helper functions started using too many default arguments,
which makes them a bit messy to use in tests, especially when we want to
add new arguments.

We change this to use method overloading instead, which makes it easier to
add new overloads without changes to existing tests.