Refactor replaceable tx publication
This commit contains very minimal logic changes: it mostly moves code
from the ReplaceableTxPublisher to two new actors (the first one to check
preconditions, the second one to encapsulate the logic for funding and
signing replaceable transactions).

This makes it easier to understand the state machine involved in each of
these steps, and will make it easier to introduce fee-bumping by reusing
these actors with minimal changes.