Improve getsentinfo availability
When using MPP, we previously inserted an entry in the payments DB only
after path-finding completed. Meanwhile,  `getsentinfo` would return an
empty array, so users cannot know that the payment is being attempted.
This could lead to users eagerly retrying and double-paying.

We now store an entry in the DB as early as possible, and remove it once
the payment is complete (unless the payment failed without sending any
HTLC, otherwise `getsentinfo` would return an empty array).

However, this doesn't entirely fix the risk of double-paying, which cannot
be completely fixed from inside eclair. External applications that make
payments on behalf of users must use the
`externalId` and check for double payments themselves before calling
eclair if they want to ensure that there's no risk of double payments.

Fixes #2055