type ChannelFlags
This has the same kind of impact as the typing of `Features`. In
particular we can now explicitly set the values in `eclair.conf`:
`eclair.channel.channel-flags.announce-channel=true`. I took the
opportunity to move this channel-related config key in its own config
section, with the goal that we move the other fields in a follow-up PR.

It also has the nice side effect of providing a nice json formatting out of
the box:

```json
"channelFlags": {
  "announceChannel": true
}
```

instead of:

```json
"channelFlags": 1
```

The `open` method of the API has been updatedi: `channelFlags` has
been replaced by a new `announceChannel` form param is available.
If user provides a `channelFlags` the API will reject the request and
warn of the API change. This is a breaking change, but I think it's
worth it because the previous param was footgunny and probably mostly
not used. This requires an update of the doc though.