Fee bumping with unsafe external input
We greatly simplify and improve the fee-bumping logic thanks to two new
bitcoind options for funding transactions.

We first use unsafe inputs to fund replaceable transactions. This ensures
that we can use all our available funds for fee bumping.

We also specify the weight of external inputs when calling bitcoind.
We were previously forced to remove our inputs before calling bitcoind
and re-add them afterwards, which forced us to have a lot of code that
was hard to read and inefficient.