Don't broadcast commit before funding locked
If a channel is force-closed before the funding transaction is deeply
confirmed, broadcasting our local commit can be a problem if the funding
tx is double spent. When that happens, the channel stays stuck in the
closing state trying to publish a commit tx with an invalid input.

If we haven't even seen the funding tx in the mempool, we have no way of
being sure that it was double spent, so we would need to keep trying
forever, which pollutes the logs with publishing errors.

Whenever we have nothing at stake and a channel closes before reaching
the normal state, we now directly go to closed.