Separate channel data from FSM data
We decouple the channel data from the channel state machine.
What states we use in our state machine should not be publicly exposed.
It is an implementation detail that may change (e.g. when we introduce a
different channel opening protocol, implemented by a separate actor).

What we now call channel data maps to the various states defined by the
lightning specification. It covers the BOLTs state machine, which can be
implemented using multiple smaller state machines in the future.

It may seem useless because we currently have almost a 1:1 mapping
between those two, but it unblocks future refactorings. It also lets us
keep transient FSM data that doesn't need to be stored in the DB along
with channel data that should be stored in the DB, which is something
we've struggled with previously.