purge expired invoice at given time interval (#2174)
When Eclair first starts, search for all unpaid expired incoming payments and purge them from the payments database. These payments take up space in the database but will no longer be paid and so should be removed to improve performance and reduce the size of the database.

Purges will be triggered every 24 hours by default but can be disabled or configured with a different interval. 
The invoice purger only looks back 15 days for expired invoices when triggered after the initial scan to reduce its impact on database performance.

Co-authored-by: Thomas HUET <81159533+thomash-acinq@users.noreply.github.com>
Co-authored-by: Bastien Teinturier <31281497+t-bast@users.noreply.github.com>