Add logs about local channel graph inclusion (#2184)
We've seen a few reports of local channels sometimes being ignored during
path-finding (e.g. #2176).

When that happens, it's currently hard to know whether the channel is in
the graph or not: knowing that would let us know if there is a bug in the
graph management code or the path-finding algorithm itself.