Update experimental trampoline feature bit
The `option_scid_alias` feature (https://github.com/lightning/bolts/pull/910)
is going to clash with the feature bit we use for trampoline payments.

Fortunately we've never advertized it, so we can simply update it to a much
higher value for now (until the final version of trampoline is specified).