Refactor blinded relay
The main reason for this refactoring is to simplify the type hierarchy
and codecs inside PaymentOnion.scala. They were very hard to understand
and the inheritance tree didn't map easily to the actual scenarios.

The goal of our codecs is to make it easy for readers to match the spec
to our code. This is why for example I chose to have private tlv codecs
that support all declared tlv fields and do the business logic of
restricting which fields are required or forbidden in custom, higher-level
codecs. This approach also lets us mark odd tlv types as forbidden if
necessary, which was impossible with the previous approach. I also chose
to relax some codec constraints to be more forward-compatible.

I added some comments to the `IncomingPaymentPacket.decrypt` function
since it's becoming quite challenging to review! I think the flow is now
still a bit complex (there are many cases to handle) but they are quite
cleanly separated so it should be easier to read.