Fix flaky replaceable tx publisher tests
There was a race condition when testing fee-bumping: we could emit the new
block event before the underlying actor registered to this event type.

This is fixed by emitting the `TransactionPublished` event after registering
to the new block events, and waiting for that in tests before publishing a
new block event.