introduce `ChannelSource` in the graph
We have two sources for channel routing parameters:
- channel updates
- routing hints in invoices

Instead of generating fake `channel_update`s from routing hints, we
define a `ChannelSource` that can be of type `Announcement` or `Hint`.

This is cleaner but also is a first step to decorrelate the identifier
of a channel in our graph, from whatever identifies the source.