make eventStream customizable in tests
This a minimal change that keep tests passing, but there is a lot
of cleanup left to do.

The issue with this technique is that we must not forget to call
`this.eventStream` instead of `context.system.eventStream` from the
`Channel` actor. However this would only affect tests since we use
the default `eventStream` in production.