Remove close() in db interfaces
It shouldn't be the responsibility of individual db classes to close
the underlying db connection because they typically share the same db
instance (postgres) or db files (sqlite). Closing should be handled in
the `Databases` level (which is already the case for postgres.

For sqlite, closing was only useful for mobile apps, which now use
lightning-kmp.

Also removed `DbFeeProvider`, which was only used by mobile apps.