Refactor router data with balance estimates
We refactor the balance estimates to be bundled with the graph in router data.
This lets the compiler find for us places where we forgot to update
balances, such as when pruning channels.

The changes to Validation.scala are probably easier to compare against
master than comparing them to the base PR, they make the change more
obvious compared to master.

We almost never use the `a max b` syntax in the codebase, we always use a
more explicit function call, so I updated it in BalanceEstimate for
consistency.

This commit doesn't contain meaningful business logic changes and should
be easy to review.