support early channel_ready as funder
As funder, if we receive a `channel_ready` from the fundee containing an
alias, we can instantly use the channel. We trust ourselves to not
double spend our own funding txs, the only thing that prevent us from
always using a zero-conf watch is that we need to have an alias from
remote.