Allow disabling bitcoind requests batching (#2301)
Eclair batches bitcoind request by default for efficiency.
However this can create latency and reliability issues when
running bitcoind on a remote machine over an unreliable
network. In those cases it's better to opt-out of batching.

We also make the threshold of missing blocks for blockchain
watchdog configurable instead of a hard-coded value of 6.