Add a channel congestion control mechanism
Channels have a limited number of HTLCs that can be in-flight at a given
time, because the commitment transaction cannot have an unbounded number
of outputs. Malicious actors can exploit this by filling our channels with
HTLCs and waiting as long as possible before failing them (also known as
a channel jamming attack).

To increase the cost of this attack, we don't let our channels be filled
with low-value HTLCs. When we already have many low-value HTLCs in-flight,
we only accept higher value HTLCs. Attackers will have to lock non-negligible
amounts to carry out the attack.