Dual funding channel confirmation (#2274)
Once we've exchanged signatures for the funding tx, we wait for it to
confirm.

Note that we don't allow mutual closing an unconfirmed channel, but that
is also the case for single-funded channels. We can improve that in the
future if necessary, but it is more efficient to double-spend an unconfirmed
channel rather than mutual close it.