Fix pruned channels coming back online
When a channel was pruned and we receive a valid channel update for it,
we need to wait until we're sure both sides of the channel are back
online: a channel that has only one side available will most likely not
be able to relay payments.

When both sides have created fresh, valid channel updates, we need to
re-request the corresponding channel announcement that we previously
deleted. We also need to request the corresponding channel updates:
if we don't receive them, we will just prune the channel again.

Fixes #2388