balance: assume send htlcs will eventually succeed
While it makes sense to assume that relayed payments will fail in the
context of balance computation (meaning that we don't earn a fee), the
opposite is true for payments sent from the local node, which will cause
the full htlc amount to be deducted from the balance.

This way we are consistent: the balance computation is pessismistic and
always assume the lowest outcome.