Receive payments for Bolt 12 invoices (#2416)
When creating a blinded route, we expose the last blinding point (that the
last node will receive).  This lets the recipient derive the corresponding
blinded private key, which they may use to sign an invoice.

We add support for generating Bolt 12 invoices and storing them in our
payments DB to then receive blinded payments.

We implement the receiving part once a blinded payment has been decrypted.
This uses the same payment flow as non-blinded payments, with slightly
different validation steps.

Note that we put a random secret in the blinded paths' path_id field
to verify that an incoming payment uses a valid blinded route generated
by us. We store that as an arbitrary byte vector to allow future changes
to this strategy.