Fix pruned channels coming back online
When a channel was pruned and we receive a valid channel update for it,
we need to wait until we're sure both sides of the channel are back
online: a channel that has only one side available will most likely not
be able to relay payments.

When both sides have created fresh, valid channel updates, we restore the
channel to our channels map and to the graph.

We remove the custom `pruned` table and instead directly use the data
from the channels table and cached data from the router.

Fixes #2388