Introduce blinded edges and hops
Since blinded routes have to be used from start to end and are somewhat
similar to Bolt 11 routing hints, we model them as an abstract single hop
during path-finding. This makes it trivial to reuse existing algorithms
without any modifications.

There is one subtlety: we previously used shortChannelIds to identify
hops (which is useful to ignore failing ones during retries and track the
amount sent to a given edge when using MPP). To minimize code changes,
we add a dummy shortChannelId to blinded edges to use as short-lived
identifiers. They don't represent "real" channels, so we make that
explicit even though it requires slightly more code.